+++ packages/injective-test-tube/src/module/mod.rs
@@ -1,6 +1,8 @@
 mod auction;
 mod authz;
 mod bank;
+mod erc20;
+mod evm;
 mod exchange;
 mod gov;
 mod helpers;
@@ -17,6 +19,11 @@ pub use test_tube_inj::module::Module;
 pub use auction::Auction;
 pub use authz::Authz;
 pub use bank::Bank;
+pub use erc20::Erc20;
+pub use evm::{
+    derive_evm_address, derive_injective_evm_address, Evm, EvmAccessListItem, EvmAccessListTx,
+    EvmCall, EvmDynamicFeeTx, EvmExecuteResponse, EvmLegacyTx, EvmQueryOptions,
+};
 pub use exchange::Exchange;
 pub use gov::Gov;
 pub use insurance::Insurance;