+++ packages/injective-test-tube/src/module/oracle.rs
@@ -229,7 +229,6 @@ mod tests {
                 params: Some(Params {
                     pyth_contract: pyth_contract.address(),
                     chainlink_verifier_proxy_contract: "".to_string(),
-                    accept_unverified_chainlink_data_streams_reports: true,
                     chainlink_data_streams_verification_gas_limit: 1000000,
                 }),
             },